Back to central location
I guess Github doesn’t like Google Docs and prefers to have the files
in a local setting.  I moved them back.

In between this, I moved the score text box beneath the buttons and
added a location message text box with the intent of the former solely
showing the score, and the latter solely giving the message.

The next step is to add the actual functions…