This is what I got...
Item functionality is still not working… I’m just so frustrated.

The disconnect is somewhere in the location prototype or array because
the item theoretically should be turned on when the user visits the
site, but the program is seeing that nothing has turned on…
I don’t know if that has to do with the syntax of the location
constructor of if its in the item constructor.