Working game :D
This is now a complete game that meets all requirements of this
project.  There are still things that I would like to add (namely
pictures), but if there is not enough time, I am happy with this
current version.

To do:
do one final sweep of code and determine if there are unnecessary lines
of code

**pizzaz